requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
streamlit==1.28.2
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
requests==2.31.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1 